Add changelog entry, add arg for bounds std, align sigma->sigma0 across likelihoods, move MAP
